#27 - moves month selection top bar to ui module
