+++ core/ui/src/main/res/values/strings.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="blank" />
+    <string name="understood_button">Entendi</string>
+    <string name="close_icon_description">Close icon</string>
+    <string name="check_icon_description">Check icon</string>
+
+    <string name="current_month_button">Mês atual</string>
+    <string name="apply_date_button">Aplicar</string>
+</resources>
